backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.34.0
pony>=0.7.19
psycopg2-binary>=2.9.10
python-jose[cryptography]>=3.3.0
bcrypt>=4.2.0
python-decouple>=3.8
pydantic>=2.10.0
email-validator>=2.3.0
